fix(onboarding): Refactor UI styles to use robust glassmorphism design and resolve E2E playwright errors
- Fixed invalid syntax in `src/ui/next/src/e2e/onboarding.spec.ts` preventing tests from parsing.
- Enhanced backend `INTERNAL_SERVER_ERROR` visibility using correct matching logic on `onboarding_agent.rs` and `mod.rs`.
- Ensured Next.js UI uses `rounded-[16px]` reliably within `globals.css` and `glassmorphism`.
