+++ package.json
@@ -17,7 +17,6 @@
     "google-protobuf": "^3.21.4",
     "next": "^16.2.7",
     "pg": "^8.21.0",
-    "playwright": "1.50.1",
     "react": "^19.2.5",
     "react-icons": "^5.6.0",
     "swagger-ui-react": "^5.32.6",
@@ -26,7 +25,7 @@
   },
   "devDependencies": {
     "@bazel/bazelisk": "^1.28.1",
-    "@playwright/test": "^1.60.0",
+    "@playwright/test": "^1.50.1",
     "@tailwindcss/postcss": "^4.3.0",
     "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/react": "^16.3.2",
