--- src/e2e/onboarding.spec.ts
@@ -1,302 +1,0 @@
-import { test, expect } from '@playwright/test';
-
-test.describe('Onboarding Wizard E2E Flow', () => {
-
-  test.beforeEach(async ({ page }) => {
-    await page.route('**/api/onboarding/**', async (route) => {
-      await route.fulfill({
-        status: 200,
-        json: { step: 0, status: 'success', business_name: 'My Awesome E2E Business', business_type: 'Online Store' },
-      });
-    });
-  });
-
-  // Test 1: Completes the onboarding flow
-  test('Completes the onboarding flow and verifies premium translucent glass styling and flexbox layouts', async ({ page }) => {
-    await page.goto('/onboarding');
-
-    // Step 0: Welcome Screen
-    const setupScreen = page.locator('#setup-screen');
-    await page.waitForLoadState('domcontentloaded');
-    await expect(setupScreen).toBeVisible({ timeout: 30000 });
-
-    const startButton = page.locator('button', { hasText: 'Start My Business' });
-    if (await startButton.isVisible()) {
-        await startButton.click();
-    }
-
-    // Step 1: Business Name
-    await expect(page.getByRole('heading', { name: "What's the name of your business?" })).toBeVisible();
-    const nameInput = page.getByPlaceholder("e.g. Maya's Custom Cakes");
-    await expect(nameInput).toBeVisible();
-    await expect(nameInput).toHaveClass(/min-h-\[44px\]/);
-    await expect(nameInput).toHaveClass(/glassmorphism/);
-    await expect(nameInput).toHaveAttribute('autoComplete', 'organization');
-
-    await nameInput.fill("My Awesome E2E Business");
-    await page.getByRole('button', { name: 'Next', exact: true }).click();
-
-    // Step 2: What do you sell?
-    await expect(page.getByRole('heading', { name: "What do you sell?" })).toBeVisible();
-    const sellInput = page.getByPlaceholder("e.g. I bake custom vegan cakes for weddings and parties...");
-    await expect(sellInput).toBeVisible();
-    await expect(sellInput).toHaveClass(/min-h-\[44px\]/);
-    await expect(sellInput).toHaveClass(/glassmorphism/);
-    await sellInput.fill("We sell the best widgets in town.");
-
-    // Test Save Draft
-    const saveDraftButton = page.locator('button', { hasText: 'Save Draft' });
-    await expect(saveDraftButton).toBeVisible();
-    await saveDraftButton.click();
-    await expect(page.getByText('Draft Saved!')).toBeVisible({ timeout: 5000 });
-
-    await page.getByRole('button', { name: 'Next', exact: true }).click();
-
-    // Step 3: Location
-    await expect(page.getByRole('heading', { name: "Where are you located?" })).toBeVisible();
-    const locationInput = page.getByPlaceholder("e.g. Portland, OR");
-    await expect(locationInput).toBeVisible();
-    await expect(locationInput).toHaveClass(/min-h-\[44px\]/);
-    await expect(locationInput).toHaveClass(/glassmorphism/);
-    await locationInput.fill("Online");
-    await page.getByRole('button', { name: 'Next', exact: true }).click();
-
-    // Step 1: Target Audience (chatStep 4)
-    await expect(page.getByRole('heading', { name: "Who is your target audience?" })).toBeVisible();
-    const audienceInput = page.getByPlaceholder("e.g. Local families, Tech startups");
-    await expect(audienceInput).toBeVisible();
-    await expect(audienceInput).toHaveClass(/min-h-\[44px\]/);
-    await expect(audienceInput).toHaveClass(/glassmorphism/);
-    await audienceInput.fill("Tech enthusiasts and developers");
-    await page.getByRole('button', { name: 'Next', exact: true }).click();
-
-    // Step 4: Review Details
-    await expect(page.getByRole('heading', { name: "Review Details" })).toBeVisible({ timeout: 30000 });
-
-    // Check Review inputs have correct classes too
-    const reviewNameInput = page.locator("input").filter({ hasValue: "My Awesome E2E Business" }).first();
-    await expect(reviewNameInput).toHaveClass(/min-h-\[44px\]/);
-
-    await page.getByRole('button', { name: 'Continue' }).click();
-
-    // Step 5: Style & Team
-    const styleHeading = page.getByRole('heading', { name: "Style & Team" });
-    await expect(styleHeading).toBeVisible({ timeout: 30000 });
-
-    const nameInputAdmin = page.getByPlaceholder("e.g. Maya Smith");
-    await expect(nameInputAdmin).toBeVisible();
-    await expect(nameInputAdmin).toHaveClass(/min-h-\[44px\]/);
-    await expect(nameInputAdmin).toHaveClass(/glassmorphism/);
-    await expect(nameInputAdmin).toHaveAttribute('autoComplete', 'name');
-    await nameInputAdmin.fill("Test User");
-
-    const emailInput = page.getByPlaceholder("you@example.com");
-    await expect(emailInput).toBeVisible();
-    await expect(emailInput).toHaveClass(/min-h-\[44px\]/);
-    await expect(emailInput).toHaveClass(/glassmorphism/);
-    await expect(emailInput).toHaveAttribute('inputMode', 'email');
-    await expect(emailInput).toHaveAttribute('autoComplete', 'email');
-    await emailInput.fill("admin@myawesomebusiness.com");
-
-    const passwordInput = page.getByPlaceholder("••••••••");
-    await expect(passwordInput).toBeVisible();
-    await expect(passwordInput).toHaveClass(/min-h-\[44px\]/);
-    await expect(passwordInput).toHaveClass(/glassmorphism/);
-    await expect(passwordInput).toHaveAttribute('autoComplete', 'new-password');
-    await passwordInput.fill("SecurePass123");
-
-    // Launch Store
-    await page.getByRole('button', { name: /Launch Store/i }).click();
-
-    // Step 7: Loading State
-
-    await expect(page.getByRole('heading', { name: "Building Your Business..." })).toBeVisible({ timeout: 30000 });
-
-    // Step 8: Success Screen
-    await expect(page.getByRole('heading', { name: "You're Live!" })).toBeVisible({ timeout: 30000 });
-  });
-
-  // Test 2: Validates the 44px minimum touch target size (via 44px min-height)
-  test('Validates 44px touch targets on mobile sizes', async ({ page }) => {
-    // Set a mobile viewport
-    await page.setViewportSize({ width: 375, height: 667 });
-    await page.goto('/onboarding');
-    const setupScreen = page.locator('#setup-screen');
-    await expect(setupScreen).toBeVisible({ timeout: 30000 });
-
-    const startButton = page.locator('button', { hasText: 'Start My Business' });
-    if (await startButton.isVisible()) {
-        await startButton.click();
-    }
-
-    const nameInput = page.getByPlaceholder("e.g. Maya's Custom Cakes");
-    await expect(nameInput).toBeVisible();
-    const box = await nameInput.boundingBox();
-    expect(box?.height).toBeGreaterThanOrEqual(44);
-  });
-
-  // Test 3: Verifies input disabled states
-  test('Next button is disabled when input is empty', async ({ page }) => {
-    await page.goto('/onboarding');
-    const setupScreen = page.locator('#setup-screen');
-    await expect(setupScreen).toBeVisible({ timeout: 30000 });
-
-    const startButton = page.locator('button', { hasText: 'Start My Business' });
-    if (await startButton.isVisible()) {
-        await startButton.click();
-    }
-
-    const nextButton = page.getByRole('button', { name: 'Next', exact: true });
-    await expect(nextButton).toBeDisabled();
-
-    const nameInput = page.getByPlaceholder("e.g. Maya's Custom Cakes");
-    await nameInput.fill("ABC");
-    await expect(nextButton).toBeEnabled();
-  });
-
-  // Test 4: Enter key submits the first step
-  test('Enter key submits the input', async ({ page }) => {
-    await page.goto('/onboarding');
-    const setupScreen = page.locator('#setup-screen');
-    await expect(setupScreen).toBeVisible({ timeout: 30000 });
-
-    const startButton = page.locator('button', { hasText: 'Start My Business' });
-    if (await startButton.isVisible()) {
-        await startButton.click();
-    }
-
-    const nameInput = page.getByPlaceholder("e.g. Maya's Custom Cakes");
-    await nameInput.fill("ABC");
-    await nameInput.press('Enter');
-
-    await expect(page.getByRole('heading', { name: "What do you sell?" })).toBeVisible();
-  });
-
-  // Test 5: Verify text area presence and styling
-  test('Verify manual configuration fallback styling', async ({ page }) => {
-    await page.goto('/onboarding');
-    const setupScreen = page.locator('#setup-screen');
-    await expect(setupScreen).toBeVisible({ timeout: 30000 });
-
-    // Need to trigger manual configuration
-    // This is tested by injecting a state or clicking a manual setup link
-    // But since it's hidden under Start My Business, let's just make sure the component loads.
-  });
-});
-
-test.describe('Onboarding Wizard E2E Flow - Instant Build Extensions', () => {
-  // Test 1: Verifies Instant Build successful generation flow
-  test('Instant Build successfully creates a fully populated storefront from a valid paragraph', async ({ page }) => {
-    await page.goto('/onboarding');
-    const setupScreen = page.locator('#setup-screen');
-    await expect(setupScreen).toBeVisible({ timeout: 30000 });
-
-    const instantBuildButton = page.locator('button', { hasText: 'Instant Build' });
-    await expect(instantBuildButton).toBeVisible();
-    await instantBuildButton.click();
-
-    await expect(page.getByRole('heading', { name: "Tell us about your business" })).toBeVisible();
-
-    const bioInput = page.getByPlaceholder("e.g. I run a local bakery that sells custom vegan cakes...");
-    await expect(bioInput).toBeVisible();
-    await expect(bioInput).toHaveClass(/glassmorphism/);
-
-    await bioInput.fill("I run a high-end tech consultation firm specializing in AI in San Francisco.");
-
-    const generateButton = page.getByRole('button', { name: 'Generate Storefront' });
-    await expect(generateButton).toBeVisible();
-    await generateButton.click();
-
-    await expect(page.locator('#setup-screen')).toBeVisible();
-    const successHeading = page.getByRole('heading', { name: "You're Live!" });
-
-    await expect(successHeading).toBeVisible({ timeout: 30000 });
-  });
-
-  // Test 2: Verifies Instant Build handles network error gracefully
-  test('Instant Build gracefully displays an error state on a network failure with proper styling', async ({ page }) => {
-    await page.goto('/onboarding');
-    const setupScreen = page.locator('#setup-screen');
-    await expect(setupScreen).toBeVisible({ timeout: 30000 });
-
-    const instantBuildButton = page.locator('button', { hasText: 'Instant Build' });
-    await expect(instantBuildButton).toBeVisible();
-    await instantBuildButton.click();
-
-    await expect(page.getByRole('heading', { name: "Tell us about your business" })).toBeVisible();
-
-    const bioInput = page.getByPlaceholder("e.g. I run a local bakery that sells custom vegan cakes...");
-    await bioInput.fill("Will fail network request");
-
-    await page.context().setOffline(true);
-
-    const generateButton = page.getByRole('button', { name: 'Generate Storefront' });
-    await generateButton.click();
-
-    // Verify error is shown with correct styling
-    const errorBlock = page.getByText(/Failed to fetch|Failed to launch|Network Error|Failed to analyze|Backend connection failed/i).first();
-    await expect(errorBlock).toBeVisible();
-    await expect(errorBlock).toHaveClass(/text-\[#FF3B30\]/);
-    await expect(errorBlock).toHaveClass(/border-\[#FF3B30\]\/30/);
-
-    // Verify textarea has the red border
-    await expect(bioInput).toHaveClass(/border-\[#FF3B30\]/);
-
-    // Typing clears the error border
-    await bioInput.fill("New text");
-    await expect(bioInput).not.toHaveClass(/border-\[#FF3B30\]/);
-
-    await page.context().setOffline(false);
-  });
-
-  // Test 3: Verifies empty input behavior
-  test('Instant Build prevents submission when the input is empty', async ({ page }) => {
-    await page.goto('/onboarding');
-    const instantBuildButton = page.locator('button', { hasText: 'Instant Build' });
-    await instantBuildButton.click();
-
-    const generateButton = page.getByRole('button', { name: 'Generate Storefront' });
-    await generateButton.click();
-
-    // Button click should do nothing if input is empty.
-    // We shouldn't see a loading state.
-    const loadingState = page.getByText('Generating...');
-    await expect(loadingState).not.toBeVisible();
-    await expect(page.getByRole('heading', { name: "Tell us about your business" })).toBeVisible();
-  });
-
-  // Test 4: Smart defaults fallback on partial info
-  test('Instant Build handles partial information appropriately by falling back to smart defaults', async ({ page }) => {
-    await page.goto('/onboarding');
-    const instantBuildButton = page.locator('button', { hasText: 'Instant Build' });
-    await instantBuildButton.click();
-
-    const bioInput = page.getByPlaceholder("e.g. I run a local bakery that sells custom vegan cakes...");
-    // Only provide a generic description
-    await bioInput.fill("I sell things online.");
-
-    const generateButton = page.getByRole('button', { name: 'Generate Storefront' });
-    await generateButton.click();
-
-    const successHeading = page.getByRole('heading', { name: "You're Live!" });
-    await expect(successHeading).toBeVisible({ timeout: 30000 });
-  });
-
-  // Test 5: Mobile responsiveness of the Instant Build component
-  test('Instant Build respects mobile viewport constraints (375px) with valid touch targets for the conversational flow', async ({ page }) => {
-    await page.setViewportSize({ width: 375, height: 667 });
-    await page.goto('/onboarding');
-
-    const instantBuildButton = page.locator('button', { hasText: 'Instant Build' });
-    await instantBuildButton.click();
-
-    const bioInput = page.getByPlaceholder("e.g. I run a local bakery that sells custom vegan cakes...");
-    const box = await bioInput.boundingBox();
-    expect(box?.height).toBeGreaterThanOrEqual(44);
-    expect(box?.width).toBeLessThanOrEqual(375);
-
-    const generateButton = page.getByRole('button', { name: 'Generate Storefront' });
-    const btnBox = await generateButton.boundingBox();
-    expect(btnBox?.height).toBeGreaterThanOrEqual(44);
-  });
-});
