+++ src/ui/next/test-results/.last-run.json
@@ -1,6 +1,15 @@
 {
   "status": "failed",
   "failedTests": [
-    "9b8cd1bc48a798eff244-8f813b7d7d2b664d3987"
+    "fe534f0825407f213faa-7f670b38d07e12094167",
+    "fe534f0825407f213faa-c9e0b48573573a126506",
+    "fe534f0825407f213faa-d4141d767b0ffd4a3f46",
+    "fe534f0825407f213faa-b06eeed05949357de80e",
+    "fe534f0825407f213faa-e00a70247cb692840bb6",
+    "fe534f0825407f213faa-33815f5242a3647bc00d",
+    "fe534f0825407f213faa-b09be83a122286bcbe0b",
+    "fe534f0825407f213faa-1366137b87eda53015fa",
+    "fe534f0825407f213faa-4093f8ddb2bd9118f657",
+    "fe534f0825407f213faa-82602c182fe911ea140b"
   ]
 }
